+++ forui/example/lib/example.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:forui/forui.dart';
 
 class Example extends StatefulWidget {
   const Example({super.key});
@@ -14,5 +15,17 @@ class _ExampleState extends State<Example> {
   }
 
   @override
-  Widget build(BuildContext context) => const Placeholder();
+  Widget build(BuildContext context) => Column(
+    children: [
+      // Material(
+      //   child: CalendarDatePicker(initialDate: DateTime.now(), firstDate: DateTime.utc(2000),
+      //     lastDate: DateTime.utc(2030), onDateChanged: (DateTime value) {  },),
+      // ),
+      FCalendar(
+        controller: FCalendarSingleValueController(),
+        start: DateTime.utc(2000),
+        end: DateTime.utc(2030),
+      ),
+    ],
+  );
 }