+++ forui/lib/src/widgets/calendar/year_month_picker.dart
@@ -93,12 +93,14 @@ final class FCalendarYearMonthPickerStyle with Diagnosticable {
             backgroundColor: colorScheme.background,
             textStyle: typography.sm.copyWith(color: colorScheme.foreground, fontWeight: FontWeight.w500),
             hoveredBackgroundColor: colorScheme.secondary,
+            focusedBorderColor: colorScheme.foreground,
             radius: const Radius.circular(8),
           ),
           disabledStyle: FCalendarEntryStyle(
             backgroundColor: colorScheme.background,
             textStyle: typography.sm
                 .copyWith(color: colorScheme.mutedForeground.withOpacity(0.5), fontWeight: FontWeight.w500),
+            focusedBorderColor: colorScheme.background,
             radius: const Radius.circular(8),
           ),
         );